@@ -11,10 +11,8 @@ license = MIT
 license_file = LICENSE
 classifiers =
     License :: OSI Approved :: MIT License
-    Programming Language :: Python :: 2
-    Programming Language :: Python :: 2.7
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.5
+    Programming Language :: Python :: 3 :: Only
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
@@ -32,10 +30,9 @@ install_requires =
     six
     toml
     virtualenv>=15.2
-    futures;python_version<"3.2"
     importlib-metadata;python_version<"3.8"
     importlib-resources;python_version<"3.7"
-python_requires = >=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*
+python_requires = >=3.6
 
 [options.entry_points]
 console_scripts =
