@@ -1,6 +1,3 @@
-from __future__ import print_function
-from __future__ import unicode_literals
-
 import os.path
 
 from pre_commit import output
@@ -12,5 +9,5 @@ def clean(store):
     for directory in (store.directory, legacy_path):
         if os.path.exists(directory):
             rmtree(directory)
-            output.write_line('Cleaned {}.'.format(directory))
+            output.write_line(f'Cleaned {directory}.')
     return 0
