@@ -16,7 +16,7 @@ def check_all_hooks_match_files(config_file):
         if hook.always_run or hook.language == 'fail':
             continue
         elif not classifier.filenames_for_hook(hook):
-            print('{} does not apply to this repository'.format(hook.id))
+            print(f'{hook.id} does not apply to this repository')
             retv = 1
 
     return retv
