@@ -1,5 +1,3 @@
-from __future__ import unicode_literals
-
 import sys
 
 import mock
@@ -14,7 +12,7 @@ from pre_commit.color import use_color
 
 @pytest.mark.parametrize(
     ('in_text', 'in_color', 'in_use_color', 'expected'), (
-        ('foo', GREEN, True, '{}foo\033[0m'.format(GREEN)),
+        ('foo', GREEN, True, f'{GREEN}foo\033[0m'),
         ('foo', GREEN, False, 'foo'),
     ),
 )
