@@ -1,11 +1,8 @@
-from __future__ import absolute_import
-from __future__ import unicode_literals
-
 import six
 
 
 def to_text(s):
-    return s if isinstance(s, six.text_type) else s.decode('UTF-8')
+    return s if isinstance(s, str) else s.decode('UTF-8')
 
 
 def to_bytes(s):
